Add special orchestrator task to beginning of queue.
I have a task that checks on the status of other tasks and launches one or more other tasks. This special orchestrator task is run after each job using the depends_on property. I want this job to run immediately and not wait on other tasks in the same queue.